requirements.txt
Django>=6.0
python-telegram-bot==20.7
requests>=2.31.0
gunicorn>=21.0.0